Move omni search helper functions to new script
