+++ client/src/app.js
@@ -17,8 +17,8 @@ import { renderSandboxFormDiv, newSandboxDiv } from '#dom/sandbox'
 import { sayerror } from '#dom/sayerror'
 import { Menu } from '#dom/menu'
 import { first_genetrack_tolist } from '#common/1stGenetk'
-import { openSandbox } from '../appdrawer/adSandbox'
 import { InputSearch } from '../dom/search.ts'
+import { findAppDrawerElements, findgenelst } from './omniSearch'
 
 /*
 
@@ -345,7 +345,8 @@ async function makeheader(app, obj, jwt) {
 		const data = [
 			{
 				title: 'Genes',
-				items: await findgenelst(app, userInput, app.selectgenome.property('value'), tip, jwt),
+				default: true,
+				items: await findgenelst(userInput, app.selectgenome.property('value'), jwt),
 				callback: gene => {
 					app.drawer.dispatch({ type: 'is_apps_btn_active', value: false })
 					tip.hide()
@@ -502,73 +503,6 @@ function make_genome_browser_btn(app, headbox, jwt) {
 function update_genome_browser_btn(app) {
 	app.genome_browser_btn.text(app.selectgenome.node().value + ' genome browser')
 	app.genome_browser_btn.datum(app.selectgenome.node().value)
-}
-
-async function findAppDrawerElements(app, input, data, tip) {
-	const re = await dofetch3(app.cardsPath + '/index.json')
-	if (re.error) throw `Problem retrieving cards index.json`
-	const userInput = input.toLowerCase()
-	const filteredElements = re.elements
-		.filter(elem => {
-			if (elem.hidden) return false
-			let searchTermFound = (elem.searchterms || []).reduce((searchTermFound, searchTerm) => {
-				if (searchTermFound) return true
-				return searchTerm.toLowerCase().includes(userInput)
-			}, false)
-			return searchTermFound || elem.name.toLowerCase().includes(userInput)
-		})
-		.sort((a, b) => a.name.localeCompare(b.name))
-
-	const opts = {
-		app: app.drawer.opts,
-		sandboxDiv: app.drawer.opts.sandboxDiv,
-		genomes: app.genomes,
-		fromApp: true
-	}
-
-	data.push(
-		{
-			title: 'Tracks and Apps',
-			items: filteredElements.filter(c => c.type == 'card'),
-			color: '#e1edf7',
-			callback: element => {
-				app.drawer.dispatch({ type: 'is_apps_btn_active', value: false })
-				tip.hide()
-				openSandbox(element, opts)
-			}
-		},
-		{
-			title: 'Datasets',
-			items: filteredElements.filter(c => c.type == 'dsButton'),
-			color: '#e5f5e4',
-			callback: element => {
-				app.drawer.dispatch({ type: 'is_apps_btn_active', value: false })
-				tip.hide()
-				openSandbox(element, opts)
-			}
-		}
-	)
-	return data
-}
-
-async function findgenelst(app, str, genome, tip, jwt) {
-	try {
-		const data = await dofetch3('/genelookup', {
-			body: {
-				input: str,
-				genome,
-				jwt
-			}
-		})
-
-		if (data.error) throw data.error
-		if (!data.hits) throw '.hits[] missing'
-		return data.hits
-	} catch (err) {
-		// err is likely "invalid character in gene name". ignore and continue
-		// if (err.stack) console.log(err.stack)
-		// throw err
-	}
 }
 
 async function findgene2paint(str, app, genomename, jwt) {
